// src/store/modules/user/types.ts
export type RoleType = '' | '*' | 'admin' | 'user';
export interface UserState {
  headImg?: string;
  name?: string;
  avatar?: string;
  job?: string;
  organization?: string;
  location?: string;
  email?: string;
  introduction?: string;
  personalWebsite?: string;
  jobName?: string;
  organizationName?: string;
  locationName?: string;
  phone?: string;
  registrationDate?: string;
  accountId?: string;
  certification?: number;
  role: RoleType;
  routers?: any[]
  permissions?: any,
  roles?: any,
  user?: any
}